Обновление модели ItemFood

+++ Items.Domain/Components/Entities/Common/ItemFood.cs
@@ -8,8 +8,7 @@
         public float Sugar { get; private set; }
         public float Energy { get; private set; }
         public int CreatedDate { get; private set; }
-        public int MinTemp { get; private set; }
-        public int MaxTemp { get; private set; }
-
+        public float MinTemp { get; private set; }
+        public float MaxTemp { get; private set; }
     }
 }
